+++ dcicutils/diff_utils.py
@@ -59,7 +59,7 @@ class DiffManager:
 
     CONSTANTS_MAP = {None: 'null', True: 'true', False: 'false'}
 
-    def unroll(self, item, normalizer=None, _omit_subscripts=False):
+    def unroll(self, item, normalizer=None, _omit_subscripts=False, _omit_empty_containers=True):
         """
         Unrolls a JSON structure (a multi-level Python dictionary) into a (flat) dictionary of dotted keys.
         The keys in the dictionary depend on the DiffManager's style attribute.
@@ -74,10 +74,10 @@ class DiffManager:
         def traverse(item, result, label):
             if normalizer:
                 item = normalizer(label=label, item=item)
-            if isinstance(item, dict):
+            if (item or _omit_empty_containers) and isinstance(item, dict):
                 for k, v in item.items():
                     traverse(v, result, self._merge_label_key(label=label, key=k))
-            elif isinstance(item, list):
+            elif (item or _omit_empty_containers) and isinstance(item, list):
                 for i, elem in enumerate(item):
                     # TODO: Would it be simpler to just omit subscripts here?
                     traverse(elem, result,
@@ -89,20 +89,25 @@ class DiffManager:
 
         return result
 
-    def diffs(self, item1, item2, include_mappings=False, normalizer=None, _omit_subscripts=False):
+    def diffs(self, item1, item2, include_mappings=False, normalizer=None,
+              _omit_subscripts=False, _omit_empty_containers=True):
         """
         Args:
             item1: a python dictionary (representing its JSON equivalent)
             item2: a python dictionary (representing its JSON equivalent)
             include_mappings (bool): a boolean indicating whether old/new mappings should be returned
+            normalizer: an optional function that normalizes the object before comparison
             _omit_subscripts: Used INTERNALLY ONLY by .patch_diffs to cause some detail to be elided.
+            _omit_empty_containers: whether to treat "foo": [] or "foo": {} the same as if it were missing
         Returns:
             a dictionary of with keys "added", "changed", "same", and "removed" containing names of keys of each kind.
             If include_mappings=True, then "old" and "new" will be among the keys.
         """
 
-        d1 = self.unroll(item1, _omit_subscripts=_omit_subscripts, normalizer=normalizer)
-        d2 = self.unroll(item2, _omit_subscripts=_omit_subscripts, normalizer=normalizer)
+        d1 = self.unroll(item1, normalizer=normalizer,
+                         _omit_subscripts=_omit_subscripts, _omit_empty_containers=_omit_empty_containers)
+        d2 = self.unroll(item2, normalizer=normalizer,
+                         _omit_subscripts=_omit_subscripts, _omit_empty_containers=_omit_empty_containers)
         return self._diffs(d1, d2, include_mappings=include_mappings)
 
     @classmethod
@@ -140,7 +145,7 @@ class DiffManager:
         """
         Returns a list of keys for would-be-affected properties if item were a patch request for a piece of JSON.
         """
-        result = self.diffs({}, item, _omit_subscripts=True)
+        result = self.diffs({}, item, _omit_subscripts=True, _omit_empty_containers=False)
         return sorted(result.get('added', []))
 
     def _maybe_sorted(self, items, for_change_type):